move monthly_variables stuff to input handler and schedule
